Changed superscenegraph creation scheme yet again:
De-coupled the shouldCreateDefaultSuperSceneGraph and createSuperSceneGraph:
delegate methods. The shouldCreateDefaultSuperSceneGraph method turns creation
of the *internal* superscenegraph on and off, and has no effect on the
createSuperSceneGraph: method. The createSuperSceneGraph: method is always
called to give the delegate the chance to do its own superscenegraph
addition/processing. Depending on whether an internal superscenegraph has been
added, createSuperSceneGraph: will be called either with the original root or
with the root of the internal superscenegraph.

This fixes a bug in the CustomEventHandler example, where we just simply always
added a light and camera (which was bogus and led to problems with files that
already contained a camera; we would have had to copy SCSceneGraph's logic
for determining whether or not we should add a camera, but that sucks, too).

This was probably the longest log message I have ever written.